Improve tests: Filesystem layout, tooling. Introduce code formatting.
