+++ requirements-release.txt
@@ -1,3 +1,5 @@
-bump2version==1.0.0
-twine==3.2.0
-keyring==21.4.0
+bump2version>=1,<2
+twine>=3.2,<4
+keyring>=21.4,<24
+isort>=5,<6
+black>=20,<22
