+++ tests/__init__.py
@@ -2,10 +2,10 @@
 # (c) 2018 The mqttwarn developers
 
 # Configuration- and function files used by the test harness
-configfile_full = 'tests/etc/full.ini'
-configfile_service_loading = 'tests/etc/service-loading.ini'
-configfile_no_functions = 'tests/etc/no-functions.ini'
-configfile_empty_functions = 'tests/etc/empty-functions.ini'
-configfile_unknown_functions = 'tests/etc/unknown-functions.ini'
-funcfile_good = 'tests/etc/functions_good.py'
-funcfile_bad = 'tests/etc/functions_bad.py'
+configfile_full = "tests/etc/full.ini"
+configfile_service_loading = "tests/etc/service-loading.ini"
+configfile_no_functions = "tests/etc/no-functions.ini"
+configfile_empty_functions = "tests/etc/empty-functions.ini"
+configfile_unknown_functions = "tests/etc/unknown-functions.ini"
+funcfile_good = "tests/etc/functions_good.py"
+funcfile_bad = "tests/etc/functions_bad.py"
