+++ tests/acme/foobar.py
@@ -1,3 +1,3 @@
 def plugin(srv, item):
     srv.logging.debug("*** MODULE=%s: service=%s, target=%s", __file__, item.service, item.target)
-    srv.logging.info('Plugin invoked.')
+    srv.logging.info("Plugin invoked.")
