+++ tests/services/test_apprise.py
@@ -1,0 +1,200 @@
+# -*- coding: utf-8 -*-
+# (c) 2021 The mqttwarn developers
+import logging
+from unittest import mock
+from unittest.mock import call
+
+from mqttwarn.model import ProcessorItem as Item
+from mqttwarn.util import load_module_from_file
+from surrogate import surrogate
+
+
+@surrogate("apprise")
+@mock.patch("apprise.Apprise", create=True)
+@mock.patch("apprise.AppriseAsset", create=True)
+def test_apprise_success(apprise_asset, apprise_mock, srv, caplog):
+
+    with caplog.at_level(logging.DEBUG):
+
+        module = load_module_from_file("mqttwarn/services/apprise.py")
+
+        item = Item(
+            config={"baseuri": "mailtos://smtp_username:smtp_password@mail.example.org"},
+            target="test",
+            addrs=["foo@example.org", "bar@example.org"],
+            title="⚽ Message title ⚽",
+            message="⚽ Notification message ⚽",
+        )
+
+        outcome = module.plugin(srv, item)
+
+        assert apprise_mock.mock_calls == [
+            call(asset=mock.ANY),
+            call().add(
+                "mailtos://smtp_username:smtp_password@mail.example.org?from=None&to=foo@example.org,bar@example.org"
+            ),
+            call().notify(body="⚽ Notification message ⚽", title="⚽ Message title ⚽"),
+            call().notify().__bool__(),
+        ]
+
+        assert outcome is True
+        assert (
+            "Sending notification to Apprise test, addresses: ['foo@example.org', 'bar@example.org']"
+            in caplog.text
+        )
+        assert "Successfully sent message using Apprise" in caplog.text
+
+
+@surrogate("apprise")
+@mock.patch("apprise.Apprise", create=True)
+@mock.patch("apprise.AppriseAsset", create=True)
+def test_apprise_failure_no_addresses(apprise_asset, apprise_mock, srv, caplog):
+
+    with caplog.at_level(logging.DEBUG):
+
+        module = load_module_from_file("mqttwarn/services/apprise.py")
+
+        item = Item(
+            config={"baseuri": "mailtos://smtp_username:smtp_password@mail.example.org"},
+            target="test",
+            addrs=[],
+            title="⚽ Message title ⚽",
+            message="⚽ Notification message ⚽",
+        )
+
+        outcome = module.plugin(srv, item)
+
+        assert apprise_mock.mock_calls == []
+
+        assert outcome is False
+        assert (
+            "Skipped sending notification to Apprise test, no addresses configured" in caplog.text
+        )
+
+
+@surrogate("apprise")
+def test_apprise_failure_notify(srv, caplog):
+
+    with caplog.at_level(logging.DEBUG):
+
+        mock_connection = mock.MagicMock()
+
+        # Make the call to `notify` signal failure.
+        def error(*args, **kwargs):
+            return False
+
+        mock_connection.notify = error
+
+        with mock.patch(
+            "apprise.Apprise", side_effect=[mock_connection], create=True
+        ) as mock_client:
+            with mock.patch("apprise.AppriseAsset", create=True) as mock_asset:
+                module = load_module_from_file("mqttwarn/services/apprise.py")
+
+                item = Item(
+                    config={"baseuri": "mailtos://smtp_username:smtp_password@mail.example.org"},
+                    target="test",
+                    addrs=["foo@example.org", "bar@example.org"],
+                    title="⚽ Message title ⚽",
+                    message="⚽ Notification message ⚽",
+                )
+
+                outcome = module.plugin(srv, item)
+
+                assert mock_client.mock_calls == [
+                    mock.call(asset=mock.ANY),
+                ]
+                assert mock_connection.mock_calls == [
+                    call.add(
+                        "mailtos://smtp_username:smtp_password@mail.example.org?from=None&to=foo@example.org,bar@example.org"
+                    ),
+                ]
+
+                assert outcome is False
+                assert (
+                    "Sending notification to Apprise test, addresses: ['foo@example.org', 'bar@example.org']"
+                    in caplog.text
+                )
+                assert "Sending message using Apprise failed" in caplog.text
+
+
+@surrogate("apprise")
+def test_apprise_error(srv, caplog):
+
+    with caplog.at_level(logging.DEBUG):
+
+        mock_connection = mock.MagicMock()
+
+        # Make the call to `notify` raise an exception.
+        def error(*args, **kwargs):
+            raise Exception("something failed")
+
+        mock_connection.notify = error
+
+        with mock.patch(
+            "apprise.Apprise", side_effect=[mock_connection], create=True
+        ) as mock_client:
+            with mock.patch("apprise.AppriseAsset", create=True) as mock_asset:
+                module = load_module_from_file("mqttwarn/services/apprise.py")
+
+                item = Item(
+                    config={"baseuri": "mailtos://smtp_username:smtp_password@mail.example.org"},
+                    target="test",
+                    addrs=["foo@example.org", "bar@example.org"],
+                    title="⚽ Message title ⚽",
+                    message="⚽ Notification message ⚽",
+                )
+
+                outcome = module.plugin(srv, item)
+
+                assert mock_client.mock_calls == [
+                    mock.call(asset=mock.ANY),
+                ]
+                assert mock_connection.mock_calls == [
+                    call.add(
+                        "mailtos://smtp_username:smtp_password@mail.example.org?from=None&to=foo@example.org,bar@example.org"
+                    ),
+                ]
+
+                assert outcome is False
+                assert (
+                    "Sending notification to Apprise test, addresses: ['foo@example.org', 'bar@example.org']"
+                    in caplog.text
+                )
+                assert "Error sending message to test: something failed" in caplog.text
+
+
+@surrogate("apprise")
+@mock.patch("apprise.Apprise", create=True)
+@mock.patch("apprise.AppriseAsset", create=True)
+def test_apprise_success_with_sender(apprise_asset, apprise_mock, srv, caplog):
+
+    with caplog.at_level(logging.DEBUG):
+
+        module = load_module_from_file("mqttwarn/services/apprise.py")
+
+        item = Item(
+            config={
+                "baseuri": "mailtos://smtp_username:smtp_password@mail.example.org",
+                "sender": "example@example.org",
+                "sender_name": "Max Mustermann",
+            },
+            target="test",
+            addrs=["foo@example.org", "bar@example.org"],
+            title="⚽ Message title ⚽",
+            message="⚽ Notification message ⚽",
+        )
+
+        outcome = module.plugin(srv, item)
+
+        assert apprise_mock.mock_calls == [
+            call(asset=mock.ANY),
+            call().add(
+                "mailtos://smtp_username:smtp_password@mail.example.org?from=example@example.org&to=foo@example.org,bar@example.org&name=Max Mustermann"
+            ),
+            call().notify(body="⚽ Notification message ⚽", title="⚽ Message title ⚽"),
+            call().notify().__bool__(),
+        ]
+
+        assert outcome is True
+        assert "Successfully sent message using Apprise" in caplog.text
